refactor(structure): reorganize project folders to feature-based structure
