+++ src/features/index.ts
@@ -1,0 +1,2 @@
+export * from './home'
+export * from './o-marcelo'
