+++ src/shared/components/cta-banner/cta-banner.tsx
@@ -1,7 +1,7 @@
 import { ArrowRightIcon } from '@radix-ui/react-icons'
 import Image from 'next/image'
 
-import { Button } from '@/components/ui'
+import { Button } from '../ui'
 
 import laptopImage from '@/assets/laptop-image.png'
