+++ src/shared/components/footer/footer.tsx
