+++ src/shared/components/footer/index.ts
