+++ src/shared/components/icons/github-icon.tsx
@@ -1,4 +1,4 @@
-import { SvgIconProps } from '@/components/icons/interfaces'
+import { SvgIconProps } from './interfaces'
 
 export const GithubIcon = (props: SvgIconProps) => {
   return (
