+++ src/shared/components/icons/index.ts
