+++ src/shared/components/icons/interfaces.ts
