+++ src/shared/components/icons/linkedin-icon.tsx
@@ -1,4 +1,4 @@
-import { SvgIconProps } from '@/components/icons/interfaces'
+import { SvgIconProps } from './interfaces'
 
 export const LinkedinIcon = (props: SvgIconProps) => {
   return (
