+++ src/shared/components/logos/index.ts
