+++ src/shared/components/logos/logo-dica-do-nerd-horizontal.tsx
