+++ src/shared/components/logos/logo-dica-do-nerd-vertical.tsx
