+++ src/shared/components/share/index.ts
