+++ src/shared/components/share/share-mobile.tsx
