+++ src/shared/components/theme/theme-toggle.tsx
@@ -1,17 +1,14 @@
 'use client'
 
-// import * as React from "react";
 import { MoonIcon, SunIcon } from '@radix-ui/react-icons'
 import { useTheme } from 'next-themes'
-
-import { Button } from '@/components/ui/button'
-
 import {
+  Button,
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
+} from '../ui'
 
 export function ThemeToggle() {
   const { setTheme } = useTheme()
