+++ src/shared/components/ui/carousel.tsx
@@ -7,7 +7,7 @@ import useEmblaCarousel, {
 } from 'embla-carousel-react'
 
 import { cn } from '@/lib/utils'
-import { Button } from '@/components/ui/button'
+import { Button } from './button'
 
 type CarouselApi = UseEmblaCarouselType[1]
 type UseCarouselParameters = Parameters<typeof useEmblaCarousel>
